+++ client/src/pages/AddPlacePage.js
@@ -22,6 +22,7 @@ import Check from '../icons/Check';
 import PropTypes from 'prop-types';
 import Marker from '../icons/Marker';
 import { fadeIn, loading } from '../components/General/Animation';
+import uploadImage from '../hooks/updloadImage';
 
 const Img = styled.img`
   width: 100%;
@@ -102,29 +103,6 @@ const CameraLabel = styled(Label)`
   border-radius: 8px;
   cursor: pointer;
 `;
-
-function uploadImage(image) {
-  return new Promise(resolve => {
-    const formData = new FormData();
-    const createDate = Date.now();
-    const xhr = new XMLHttpRequest();
-    const url = `https://api.cloudinary.com/v1_1/befamily/upload`;
-    xhr.open('POST', url, true);
-    xhr.setRequestHeader('X-Requested-With', 'XMLHttpRequest');
-    formData.append('upload_preset', 'nd1vsnsz');
-    formData.append('file', image, createDate);
-    formData.append('name', createDate);
-    formData.append('public_id,', createDate);
-    xhr.send(formData);
-    xhr.onreadystatechange = function() {
-      if (xhr.readyState === 4 && xhr.status === 200) {
-        const response = JSON.parse(xhr.responseText);
-        const url = response.secure_url;
-        resolve(url);
-      }
-    };
-  });
-}
 
 export default function AddPlacePage({ onAddPlace }) {
   const [place, setPlace] = React.useState({
